created users app and related files, move models
