+++ smse_onboarding_main_project/users/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from . import views
+
+app_name = 'users'
+
+urlpatterns = [
+    path('profile/', views.profile, name='profile'),
+    #TODO:adding urls here
+]
